Add support to mod_oreka for native codec recording
Native codec recording sends the stream to orkaudio in the same
format as the session being recorded.

You need to specify 'native-streams' as 'yes' in the oreka.conf.xml
configuration file or the variable 'oreka_native_streams=yes' in the
session executing the oreka_record application.

Native codec recording does not work if you want to also mux the streams
with the 'mux-all-streams' option or session variable 'oreka_mux_streams'.