[mod_oreka]: Do not inherit the global local port for session config
Using the global port for session configuration can lead easily to
conflicts. The default should be zero to let the OS pick a port. Only
if the user specifies a specific local port should we use it for the
session and it's up to them to avoid conflicts in such cases.